API

The exported symbols from this package define its interface. Some symbols from other packages are re-exported for convenience. Fields of objects with composite types should not be accessed directly; the internals of any given structure may change at any time and this would not be considered a breaking change.

Fitting a model

BetaRegression.BetaRegressionModel — Type

BetaRegressionModel{T,L1,L2,V,M} <: RegressionModel

Type representing a regression model for beta-distributed response values in the open interval (0, 1), as described by Ferrari and Cribari-Neto (2004).

The mean response is linked to the linear predictor by a link function with type L1 <: Link01, i.e. the link must map $(0, 1) \mapsto \mathbb{R}$ and use the GLM package's interface for link functions. While there is no canonical link function for the beta regression model as there is for GLMs, logit is the most common choice.

The precision is transformed by a link function with type L2 <: Link which should map $\mathbb{R} \mapsto \mathbb{R}$ or, ideally, $(0, \infty) \mapsto \mathbb{R}$ because the precision must be positive. The most common choices are the identity, log, and square root links.

StatsAPI.fit — Function

fit(BetaRegressionModel, formula, data, link=LogitLink(), precisionlink=IdentityLink();
    kwargs...)

Fit a BetaRegressionModel to the given table data, which may be any Tables.jl-compatible table (e.g. a DataFrame), using the given formula, which can be constructed using @formula. In this method, the response and model matrix are determined from the formula and table. It is also possible to provide them explicitly.

fit(BetaRegressionModel, X::AbstractMatrix, y::AbstractVector, link=LogitLink(),
    precisionlink=IdentityLink(); kwargs...)

Fit a beta regression model using the provided model matrix X and response vector y. In both of these methods, a link function may be provided, otherwise the default logit link is used. Similarly, a link for the precision may be provided, otherwise the default identity link is used.

Keyword Arguments

weights: A vector of weights or nothing (default). Currently only nothing is accepted.

offset: An offset vector to be added to the linear predictor or nothing (default).

maxiter: Maximum number of Fisher scoring iterations to use when fitting. Default is 100.

atol: Absolute tolerance to use when checking for model convergence. Default is sqrt(eps(T)) where T is the type of the estimates.

rtol: Relative tolerance to use when checking for convergence. Default is the Base default relative tolerance for T.

Note

If you experience convergence issues, you may consider trying a different link for the precision; LogLink() is a common choice. Increasing the maximum number of iterations may also be beneficial, especially when working with Float32.

StatsAPI.fit! — Function

fit!(b::BetaRegressionModel{T}; maxiter=100, atol=sqrt(eps(T)), rtol=Base.rtoldefault(T))

Fit the given BetaRegressionModel, updating its values in-place. If model convergence is achieved, b is returned, otherwise a ConvergenceException is thrown.

Fitting the model consists of computing the maximum likelihood estimates for the coefficients and precision parameter via Fisher scoring with analytic derivatives. The model is determined to have converged when the score vector, i.e. the vector of first partial derivatives of the log likelihood with respect to the parameters, is approximately zero. This is determined by isapprox using the specified atol and rtol. maxiter dictates the maximum number of Fisher scoring iterations.

Properties of a model

The following common functions are extended for beta regression models:

Link: The model's link function

aic: Akaike's information criterion for the model

aicc: Model AIC corrected for small sample sizes

bic: Bayesian information criterion for the model

coef: The vector $\boldsymbol{\beta}$ of regression coefficients

coefnames: Names of the coefficients (for models fit with a formula and table)

coeftable: Table of coefficient names, values, standard errors, z-values, and p-values

confint: Confidence intervals for the coefficient estimates

deviance: Model deviance

devresid: Vector of deviance residuals

dof: Degrees of freedom

dof_residual: Residual degrees of freedom

fitted: The vector $\hat{\mathbf{y}}$ of fitted values from the model

informationmatrix: Expected or observed Fisher information

linearpredictor: The linear predictor vector $\boldsymbol{\eta}$

Link: Link function $g$ used for the mean

loglikelihood: Model log likelihood

modelmatrix: The model matrix $\mathbf{X}$

nobs: Number of observations used to fit the model

offset: Model offset, empty if the model was not fit with an offset

params: All parameters from the model, including both $\boldsymbol{\beta}$ and $\phi$

precision: The estimated precision parameter $\phi$ on its natural scale

precisionlink: Link function $h$ used for the precision parameter

predict: Predict new response values given new observations

r2/r²: Pseudo $R^2$

residuals: Vector of residuals

response: The response vector $\boldsymbol{y}$

responsename: Name of the response variable (for models fit with a formula and table)

score: Score vector

stderror: Standard errors of the coefficient and precision parameter estimates

vcov: Variance-covariance matrix

weights: Model weights, empty if the model was not fit with weights

Note that for a model with $p$ independent variables, the information and variance-covariance matrices will have $p + 1$ rows and columns, the last of which corresponds to the precision term. However, coef does not include the precision term and will have length $p$.

Developer documentation

This section documents some functions that are not user facing (and are thus not exported) and may be removed at any time. They're included here for the benefit of anyone looking to contribute to the package and wondering how certain internals work.

BetaRegression.dmueta — Function

dmueta(link::Link, η)

Return the second derivative of linkinv, $d²μ/dη²$, of the link function link evaluated at the linear predictor value η. A method of this function must be defined for a particular link function in order to compute the observed information matrix.

BetaRegression.initialize! — Function

initialize!(b::BetaRegressionModel)

Initialize the given BetaRegressionModel by computing starting points for the parameter estimates and return the updated model object. The initial estimates are based on those from a linear regression model with the same model matrix as b but with linkfun.(Link(b), response(b)) as the response.